requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
python-decouple
celery
redis
python-telegram-bot==13.15